+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### 🚀 New
 
-* The `dirname` in an `ImageNamer` now accepts the `{sjd}` placeholder which will be filled out with the SDSS-style MJD (as returned by sdsstools `get_sjd()`).
+* [#26](https://github.com/sdss/basecam/issues/26) The `dirname` in an `ImageNamer` now accepts the `{sjd}` placeholder which will be filled out with the SDSS-style MJD (as returned by sdsstools `get_sjd()`).
 
 ### 🔧 Fixed